+++ CHANGES.rst
@@ -23,7 +23,7 @@ upgrading your version of coverage.py.
 Unreleased
 ----------
 
-Nothing yet.
+- Fix: ``patch = execv`` also inherits the entire coverage configuration now.
 
 
 .. start-releases